requirements.txt
connexion==2.2.0
python-jose==3.0.1
six==1.9
Flask==0.10.1
praw==6.4.0
sqlalchemy==1.3.15
swagger_ui_bundle